+++ R/refit.R
@@ -23,12 +23,11 @@
 #' @rdname refit
 #' @export
 refit.mdl_df <- function(object, new_data, ...){
-  mdls <- syms(mable_vars(object))
-  new_data <- bind_new_data(object, new_data)
-  
-  object %>% 
-    dplyr::mutate_at(vars(!!!mdls),
-                     refit, new_data[["new_data"]], ...)
+  mdls <- mable_vars(object)
+  object <- bind_new_data(object, new_data)
+  new_data <- object[["new_data"]]
+  object[["new_data"]] <- NULL
+  dplyr::mutate_at(object, vars(!!!mdls), refit, new_data, ...)
 }
 
 #' @export